# backend/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Exchange Rate API
  version: 1.0.0
  description: REST API for stored Korea Eximbank exchange-rate data, favorites, calculations, and portfolio simulation.
servers:
  - url: http://localhost:8080
paths:
...
  /api/portfolio/transactions:
    get:
      summary: Search portfolio transactions
      description: Searches portfolio transactions by curUnit, transactionType, and date range.
      parameters:
        - $ref: '#/components/parameters/PortfolioIdQuery'
        - name: curUnit
          in: query
          required: false
          schema:
            type: string
        - name: transactionType
          in: query
          required: false
          schema:
            type: string
            enum: [BUY, SELL, DEPOSIT]
        - name: fromDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paged portfolio transaction result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioTransactionPage'
  /api/portfolio/transactions/{transactionId}:
    delete:
      summary: Delete portfolio transaction
      parameters:
        - $ref: '#/components/parameters/UserKeyHeader'
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserKeyHeader:
      name: X-USER-KEY
      in: header
      required: false
      description: User key for MVP portfolio data. Defaults to demo-user when omitted.
      schema:
        type: string
        default: demo-user
    PortfolioIdQuery:
      name: portfolioId
      in: query
      required: false
      description: Selected portfolio account id. When omitted, the first portfolio is used.
      schema:
        type: integer
        format: int64
...
  schemas:
...
    PortfolioTransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        curUnit:
          type: string
          nullable: true
        normalizedCurUnit:
          type: string
          nullable: true
        curName:
          type: string
        transactionType:
          type: string
          enum:
            - BUY
            - SELL
            - DEPOSIT
        transactionDate:
          type: string
          format: date
        foreignAmount:
          type: number
          nullable: true
        exchangeRate:
          type: number
          nullable: true
        unitFactor:
          type: number
          nullable: true
        amount:
          type: number
        memo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PortfolioTransactionPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioTransactionResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
...